+++ tardis/montecarlo/montecarlo_numba/continuum/r_packet_transport_continuum.py
@@ -8,7 +8,7 @@ from tardis.transport.calculate_distances import (
     calculate_distance_line,
 )
 from tardis.montecarlo.montecarlo_numba.estimators import update_line_estimators
-from tardis.montecarlo.montecarlo_numba.frame_transformations import (
+from tardis.transport.frame_transformations import (
     get_doppler_factor,
 )
