+++ tardis/transport/r_packet_transport.py
@@ -12,7 +12,7 @@ from tardis.montecarlo.montecarlo_numba.estimators import (
     update_line_estimators,
     set_estimators,
 )
-from tardis.montecarlo.montecarlo_numba.frame_transformations import (
+from tardis.transport.frame_transformations import (
     get_doppler_factor,
 )
 from tardis.montecarlo.montecarlo_numba.numba_config import (
